requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==42.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.11
django-allauth==0.61.1
django-crispy-forms==2.1
django-phonenumber-field==7.3.0
gunicorn==21.2.0
idna==3.6
oauthlib==3.2.2
packaging==24.0
phonenumbers==8.13.32
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.4.0
slugify==0.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
